refactor(validation): unify schema cache
