+++ docs/imednet.validation.rst
@@ -4,10 +4,10 @@ imednet.validation package
 Submodules
 ----------
 
-imednet.validation.async\_schema module
----------------------------------------
+imednet.validation.cache module
+-------------------------------
 
-.. automodule:: imednet.validation.async_schema
+.. automodule:: imednet.validation.cache
    :members:
    :undoc-members:
    :show-inheritance:
