+++ docs/schema_validation.rst
@@ -1,8 +1,8 @@
 Schema Validation Flow
 ======================
 
-The SDK uses :class:`~imednet.validation.schema.SchemaValidator` and
-:class:`~imednet.validation.schema.SchemaCache` to verify record data before
+The SDK uses :class:`~imednet.validation.cache.SchemaValidator` and
+:class:`~imednet.validation.cache.SchemaCache` to verify record data before
 submitting it to the API. The validator checks that all variables exist,
 required fields are present, and that values match the expected types.
 If the variable metadata for a form is missing, ``SchemaValidator`` automatically
@@ -10,9 +10,9 @@ loads it from the API using :class:`~imednet.endpoints.variables.VariablesEndpoi
 Any problems raise :class:`~imednet.core.exceptions.ValidationError` before the
 record is sent to the server.
 
-``SchemaValidator`` and ``AsyncSchemaValidator`` share their core validation
-steps through a common mixin. The only difference is whether the cached
-metadata is refreshed using synchronous or asynchronous API calls.
+``SchemaValidator`` works with both synchronous and asynchronous SDK clients.
+It refreshes cached metadata using the appropriate API calls depending on the
+client type.
 
 The diagram below outlines the main steps.
 
@@ -33,11 +33,11 @@ The diagram below outlines the main steps.
    I -- No --> K[raise ValidationError]
 
 Record payloads can also be validated asynchronously. Use
-``AsyncSchemaValidator.validate_batch`` with ``AsyncImednetSDK`` before
+``SchemaValidator.validate_batch`` with ``AsyncImednetSDK`` before
 submitting records::
 
     async with AsyncImednetSDK() as sdk:
-        validator = AsyncSchemaValidator(sdk)
+        validator = SchemaValidator(sdk)
         await validator.validate_batch(study_key, records)
         await sdk.records.async_create(study_key, records, schema=validator.schema)
 
@@ -50,7 +50,7 @@ functions with ``SchemaCache`` to validate payloads locally::
 
     from types import SimpleNamespace
     from imednet.testing import fake_data
-    from imednet.validation.schema import SchemaCache
+    from imednet.validation.cache import SchemaCache
 
     forms = fake_data.fake_forms_for_cache(1, study_key="S")
     variables = fake_data.fake_variables_for_cache(forms, vars_per_form=2,
