+++ tests/unit/test_utils_schema.py
@@ -4,7 +4,7 @@ import pytest
 from imednet.core.exceptions import UnknownVariableTypeError, ValidationError
 from imednet.models.forms import Form
 from imednet.models.variables import Variable
-from imednet.validation.schema import (
+from imednet.validation.cache import (
     SchemaCache,
     SchemaValidator,
     _check_type,
