+++ tests/utils/test_lazy_attrs.py
@@ -1,5 +1,5 @@
 from imednet.utils import records_to_dataframe
-from imednet.validation.schema import SchemaCache
+from imednet.validation.cache import SchemaCache
 
 
 def test_lazy_attrs_available() -> None:
